Updated the windows file!
Now the file saves properly, unlike the Linux one; I had to learn how to do that part. I will be removing the load file option since when this game is executed, it autoloads. The Linux version will be updated tomorrow. Thanks for viewing the description if you have. Have a lovely day!